+++ abc-bank/Transaction.cs
@@ -4,12 +4,12 @@ namespace abc_bank
 {
     public class Transaction
     {
-        public readonly double _amount;
-        public double Amount => _amount;
+        public readonly decimal _amount;
+        public decimal Amount => _amount;
 
         private DateTime transactionDate;
 
-        public Transaction(double amount) 
+        public Transaction(decimal amount) 
         {
             _amount = amount;
             transactionDate = DateProvider.getInstance().Now();
